LayoutTests/platform/glib/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/operator-dictionary-stretchy-003-expected.txt


PASS Operator dictionary chunk 1 - stretchy
PASS Operator dictionary chunk 2 - stretchy
PASS Operator dictionary chunk 3 - stretchy
PASS Operator dictionary chunk 4 - stretchy
FAIL Operator dictionary chunk 5 - stretchy assert_approx_equals: Stretchy property for 𞻱 postfix should be 'true' expected 127.75 +/- 1 but got 77.75
stretchy for "𞻱" (postfix):
  
𞻱
 VS
  
𞻱
